+++ tests/utils/test_validation.py
@@ -1,9 +1,12 @@
 import pytest
 
+from jsonschema import ValidationError
+
 from rasa.constants import DOMAIN_SCHEMA_FILE, CONFIG_SCHEMA_FILE
 
 import rasa.utils.validation as validation_utils
 import rasa.utils.io as io_utils
+import rasa.nlu.schemas.data_schema as schema
 
 
 @pytest.mark.parametrize(
@@ -31,3 +34,74 @@ def test_validate_yaml_schema(file, schema):
 def test_validate_yaml_schema_raise_exception(file, schema):
     with pytest.raises(validation_utils.InvalidYamlFileError):
         validation_utils.validate_yaml_schema(io_utils.read_file(file), schema)
+
+
+def test_example_training_data_is_valid():
+    demo_json = "data/examples/rasa/demo-rasa.json"
+    data = io_utils.read_json_file(demo_json)
+    validation_utils.validate_training_data(data, schema.rasa_nlu_data_schema())
+
+
+@pytest.mark.parametrize(
+    "invalid_data",
+    [
+        {"wrong_top_level": []},
+        ["this is not a toplevel dict"],
+        {
+            "rasa_nlu_data": {
+                "common_examples": [{"intent": "some example without text"}]
+            }
+        },
+        {
+            "rasa_nlu_data": {
+                "common_examples": [
+                    {
+                        "text": "mytext",
+                        "entities": [{"start": "INVALID", "end": 0, "entity": "x"}],
+                    }
+                ]
+            }
+        },
+    ],
+)
+def test_validation_is_throwing_exceptions(invalid_data):
+    with pytest.raises(ValidationError):
+        validation_utils.validate_training_data(
+            invalid_data, schema.rasa_nlu_data_schema()
+        )
+
+
+def test_url_data_format():
+    data = """
+    {
+      "rasa_nlu_data": {
+        "entity_synonyms": [
+          {
+            "value": "nyc",
+            "synonyms": ["New York City", "nyc", "the big apple"]
+          }
+        ],
+        "common_examples" : [
+          {
+            "text": "show me flights to New York City",
+            "intent": "unk",
+            "entities": [
+              {
+                "entity": "destination",
+                "start": 19,
+                "end": 32,
+                "value": "NYC"
+              }
+            ]
+          }
+        ]
+      }
+    }"""
+    fname = io_utils.create_temporary_file(
+        data.encode(io_utils.DEFAULT_ENCODING),
+        suffix="_tmp_training_data.json",
+        mode="w+b",
+    )
+    data = io_utils.read_json_file(fname)
+    assert data is not None
+    validation_utils.validate_training_data(data, schema.rasa_nlu_data_schema())
